Adding credits to original author of this script
This script is essentially a bunch of modifications to a script that could be found in https://github.com/FelixDesrochers/Numerov. I essentially built a few functions on top of what Felix had done and changed the syntax in a way that would fit QPy's. However, it is still necessary to give Felix credit.